+++ example/lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:formstack/formstack.dart';
-import 'package:flutter/cupertino.dart';
 
 void main() {
   runApp(const MainApp());
@@ -23,8 +22,7 @@ class MainApp extends StatelessWidget {
         theme: ThemeData(
           useMaterial3: true,
           colorScheme: ColorScheme.fromSwatch(
-              primarySwatch: Colors.blue,
-              backgroundColor: CupertinoColors.white),
+              primarySwatch: Colors.blue, backgroundColor: Colors.white),
         ),
         home: const HomeScreen(),
         debugShowCheckedModeBanner: false);