+++ lib/src/relevant/dynamic_relevant_condition.dart
@@ -1,6 +1,10 @@
 import 'package:formstack/src/relevant/relevant_condition.dart';
 
+/// Condition Relevant for dynamic check
 class DynamicConditionalRelevant extends RelevantCondition {
+  ///
+  /// DynamicConditionalRelevant(identifier: GenericIdentifier(id:"DEMO"),isValidCallBack(resut){})
+  ///
   DynamicConditionalRelevant(
       {required super.identifier,
       super.formName,
@@ -8,6 +12,10 @@ class DynamicConditionalRelevant extends RelevantCondition {
       : _isValidCallBack = isValidCallBack;
 
   final Function(dynamic)? _isValidCallBack;
+
+  ///
+  /// Check the result is valid or not
+  ///
   @override
   bool isValid(dynamic result) {
     return _isValidCallBack!.call(result);