+++ lib/src/ui/views/display_step_view.dart
@@ -1,11 +1,8 @@
-import 'dart:io';
-import 'dart:math';
-
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:formstack/src/step/display_step.dart';
 import 'package:formstack/src/ui/views/base_step_view.dart';
-import 'package:webviewx/webviewx.dart';
+import 'package:formstack/src/ui/views/tiles_view.dart';
+import 'package:formstack/src/ui/views/web_view.dart';
 
 // ignore: must_be_immutable
 class DisplayStepView extends BaseStepView<DisplayStep> {
@@ -14,37 +11,12 @@ class DisplayStepView extends BaseStepView<DisplayStep> {
 
   @override
   Widget? buildWInputWidget(BuildContext context, DisplayStep formStep) {
-    if (kIsWeb || Platform.isAndroid || Platform.isIOS) {
-      return WebViewX(
-        key: const ValueKey('webviewx'),
-        initialSourceType: SourceType.url,
-        height: min(
-            MediaQuery.of(context).size.height - kToolbarHeight * 0.8, 1024),
-        width: min(MediaQuery.of(context).size.width * 0.8, 1024),
-        onWebViewCreated: (controller) async {
-          controller.loadContent(formStep.url, SourceType.url);
-        },
-        onPageStarted: (src) =>
-            debugPrint('A new page has started loading: $src\n'),
-        onPageFinished: (src) =>
-            debugPrint('The page has finished loading: $src\n'),
-        jsContent: const {},
-        dartCallBacks: const {},
-        webSpecificParams: const WebSpecificParams(
-          printDebugInfo: true,
-        ),
-        mobileSpecificParams: const MobileSpecificParams(
-          androidEnableHybridComposition: true,
-        ),
-        navigationDelegate: (navigation) {
-          debugPrint(navigation.content.sourceType.toString());
-          return NavigationDecision.navigate;
-        },
-      );
-    } else {
-      return ConstrainedBox(
-          constraints: const BoxConstraints(maxWidth: 500.0),
-          child: const Text("Unsupported platfom"));
+    switch (formStep.displayStepType) {
+      case DisplayStepType.listTile:
+        return ListTitlesView.buildView(context, formStep);
+      case DisplayStepType.web:
+      default:
+        return WebView.buildView(context, formStep);
     }
   }