+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.4.0
-* NesterQuestionStep creted to display mutiple questions in single screen.
+*  NestedQuestionStep -> Renamed to -> NestedStep
+*  To Shoe nested step in single view
+## 0.4.0
+* NestedQuestionStep creted to display mutiple questions in single screen.
 * Documents updated and code optimisation.
 ## 0.2.10
 * Documents updated and code optimisation.